Renamed Rectangle file, fixed forward definition for Rectangle struct
